refactor: move mcp server logic to independent src/ folder and dependencies
