+++ src/mcp-server-plexosdb/src/mcp_server_plexosdb/__init__.py
@@ -1,1 +1,9 @@
 """Standalone package exposing the plexosdb MCP server CLI."""
+
+from mcp_server_plexosdb.server import (
+    MCPServerState,
+    build_mcp_server,
+    main,
+)
+
+__all__ = ("MCPServerState", "build_mcp_server", "main")
