+++ src/mcp-server-plexosdb/src/mcp_server_plexosdb/__main__.py
@@ -4,11 +4,11 @@ from __future__ import annotations
 
 import sys
 
-from plexosdb.mcp_server import main as _main
+from mcp_server_plexosdb import main as _main
 
 
 def main() -> None:
-    """Delegate execution to the MCP server implementation in plexosdb."""
+    """Delegate execution to the MCP server implementation."""
     _main(sys.argv[1:])
